+++ src/main/java/de/thomas_oster/liblasercut/Raster3dPart.java
@@ -121,7 +121,7 @@ public class Raster3dPart extends RasterizableJobPart
     // x = 255 -> y = 0
     // 
     // x = 0  ->  y = <max>  ->  y = m*0 + c  ->  c = <max>
-    float c = (float) power.getPower();
+    float c = power.getPower();
     
     // x = 255  ->  y = 0  ->  y = m*255 + <max>  ->  0 = m*255 + <max>
     // ->  -<max> = m*255  -> -<max>/255 = m
